+++ src/rust/bitbox02-rust-c/src/async_usb.rs
@@ -61,6 +61,6 @@ pub extern "C" fn rust_async_usb_on_request_hww(usb_in: util::bytes::Bytes) {
 }
 
 #[unsafe(no_mangle)]
-pub extern "C" fn rust_async_usb_cancel() {
+pub extern "C" fn rust_async_usb_cancel() -> bool {
     bitbox02_rust::async_usb::cancel()
 }